# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sales Outreach Field Notes API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /businesses/{id}/visits:
    get:
      operationId: listVisitsForBusiness
      tags: [visits]
      summary: List visits for a business
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of visits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Visit"
  /visits:
    get:
      operationId: listVisits
      tags: [visits]
      summary: List all visits
      responses:
        "200":
          description: List of all visits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Visit"
    post:
      operationId: createVisit
      tags: [visits]
      summary: Create a visit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVisitBody"
      responses:
        "201":
          description: Created visit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Visit"
  /visits/{id}:
    get:
      operationId: getVisit
      tags: [visits]
      summary: Get a visit with all notes and media
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Visit with notes and media
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VisitDetail"
        "404":
          description: Not found
    put:
      operationId: updateVisit
      tags: [visits]
      summary: Update a visit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVisitBody"
      responses:
        "200":
          description: Updated visit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Visit"
    delete:
      operationId: deleteVisit
      tags: [visits]
      summary: Delete a visit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Visit:
      type: object
      properties:
        id:
          type: integer
        businessId:
          type: integer
        businessName:
          type: string
        visitedAt:
          type: string
          format: date-time
        outcome:
          type: string
          enum: [positive, neutral, negative, follow_up_needed]
        contactName:
          type: string
        contactPhone:
          type: string
        nextActionDate:
          type: string
          format: date-time
        noteCount:
          type: integer
        mediaCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - businessId
        - visitedAt
        - outcome
        - noteCount
        - mediaCount
        - createdAt
    VisitDetail:
      type: object
      properties:
        id:
          type: integer
        businessId:
          type: integer
        businessName:
          type: string
        visitedAt:
          type: string
          format: date-time
        outcome:
          type: string
          enum: [positive, neutral, negative, follow_up_needed]
        contactName:
          type: string
        contactPhone:
          type: string
        nextActionDate:
          type: string
          format: date-time
        notes:
          type: array
          items:
            $ref: "#/components/schemas/Note"
        media:
          type: array
          items:
            $ref: "#/components/schemas/Media"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - businessId
        - visitedAt
        - outcome
        - notes
        - media
        - createdAt
    CreateVisitBody:
      type: object
      properties:
        businessId:
          type: integer
        visitedAt:
          type: string
          format: date-time
        outcome:
          type: string
          enum: [positive, neutral, negative, follow_up_needed]
        contactName:
          type: string
        contactPhone:
          type: string
        nextActionDate:
          type: string
          format: date-time
      required:
        - businessId
        - visitedAt
        - outcome
    UpdateVisitBody:
      type: object
      properties:
        visitedAt:
          type: string
          format: date-time
        outcome:
          type: string
          enum: [positive, neutral, negative, follow_up_needed]
        contactName:
          type: string
        contactPhone:
          type: string
        nextActionDate:
          type: string
          format: date-time
    Note:
      type: object
      properties:
        id:
          type: integer
        visitId:
          type: integer
        type:
          type: string
          enum: [text, voice]
        content:
          type: string
        audioUrl:
          type: string
        durationSeconds:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - visitId
        - type
        - createdAt
...
    Media:
      type: object
      properties:
        id:
          type: integer
        visitId:
          type: integer
        type:
          type: string
          enum: [image, voice_note, interview, document]
        url:
          type: string
        filename:
          type: string
        caption:
          type: string
        mimeType:
          type: string
        sizeBytes:
          type: integer
        transcript:
          type: string
        transcriptionStatus:
          type: string
          enum: [none, pending, processing, done, error]
        transcriptionError:
          type: string
        aiStructured:
          $ref: "#/components/schemas/StructuredNote"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - visitId
        - type
        - url
        - filename
        - transcriptionStatus
        - createdAt
    StructuredNote:
      type: object
      properties:
        summary:
          type: string
        interestLevel:
          type: string
          enum: [hot, warm, cool, cold, unknown]
        objections:
          type: array
          items:
            type: string
        followUpItems:
          type: array
          items:
            type: string
        contactInfo:
          type: string
        nextStep:
          type: string
...