# lib/api-zod/src/generated/types/media.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sales Outreach Field Notes API
 * OpenAPI spec version: 0.1.0
 */
import type { MediaTranscriptionStatus } from "./mediaTranscriptionStatus";
import type { MediaType } from "./mediaType";
import type { StructuredNote } from "./structuredNote";

export interface Media {
  id: number;
  visitId: number;
  type: MediaType;
  url: string;
  filename: string;
  caption?: string;
  mimeType?: string;
  sizeBytes?: number;
  transcript?: string;
  transcriptionStatus: MediaTranscriptionStatus;
  transcriptionError?: string;
  aiStructured?: StructuredNote;
  createdAt: Date;
}