# lib/api-zod/src/generated/types/structuredNote.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sales Outreach Field Notes API
 * OpenAPI spec version: 0.1.0
 */
import type { StructuredNoteInterestLevel } from "./structuredNoteInterestLevel";

export interface StructuredNote {
  summary?: string;
  interestLevel?: StructuredNoteInterestLevel;
  objections?: string[];
  followUpItems?: string[];
  contactInfo?: string;
  nextStep?: string;
}